Version v4.0.31-rc-1

+++ FINT.Model.Resource.Arkiv/Noark/KlasseResource.cs
@@ -11,7 +11,7 @@ namespace FINT.Model.Arkiv.Noark
 
     
         public string KlasseId { get; set; }
-        public int? Rekkefolge { get; set; }
+        public long? Rekkefolge { get; set; }
         public SkjermingResource Skjerming { get; set; }
         public string Tittel { get; set; }
