+++ FINT.Model.Utdanning/Vurdering/Fravarsprosent.cs
@@ -7,9 +7,9 @@ namespace FINT.Model.Utdanning.Vurdering
 {
 	public class Fravarsprosent {
 	
-		public int Fravarstimer { get; set; }
-		public int Prosent { get; set; }
-		public int Undervisningstimer { get; set; }
+		public long Fravarstimer { get; set; }
+		public long Prosent { get; set; }
+		public long Undervisningstimer { get; set; }
 		
 	}
 }
